Reorganize purgers
