+++ packages/purge-cache-plugin/src/index.ts
@@ -8,6 +8,6 @@ export {
   type PurgerAction,
 } from './types.js';
 
-export { getCloudflarePurgerAction } from './purger/cloudflare.js';
-export { getNextjsPurgerAction } from './purger/nextjs.js';
-export { getHttpPurgerAction } from './purger/http.js';
+export { getCloudflarePurgerAction } from './purgers/cloudflare.js';
+export { getNextjsPurgerAction } from './purgers/nextjs.js';
+export { getHttpPurgerAction } from './purgers/http.js';
